Spring MVC controller configured and added
I have changed the project structure so it can reflect the structure of a Spring MVC application. Tested the configuration on Jetty (local environment) and the mapping seems to be working.
As mentioned in various sources the .jsp is the preferred format so changed all the .html files to .jsp
The next step will be to actually be able to connect to the core and get the required data. Also some cosmetic changes with the look of the GUI so it will look nicer.