Configuration changes with the Spring MVC
I have added some more dependencies to the poms in the child modules (jetty, for testing purposes). I have added the database configuration in the web module so now it will start with the access to the db.
Next step will be to actually try to get the data from the db and show it on the site (code in jsp to iterate through the results from dv has been addded).