Changes in the GUI
We are now able to actually take the data from the DB and show in on the web page. It is working with my local db with little data, but it handles it properly and after actual
data from the parser will be added to the DB we will be able to show it on the page. Currently this branch will be waiting for the parser to be able to save data and after that and I will
confirm it is shown properly this branch can be merged into master.