Major changes to the GUI
So with the usage of bootstrap changed the look of the gui. Now we can see the table with all the books (for now only for Empik) and the description can be shown/hidden when the link is clicked.
There are some missing places for Tags and Genre, but this is caused by the information sent from the database, changing that will be done in another branch.
As for the next commit for this branch, I will add the filtering on the table as well as paging of the results.