Cache creation
I have implemented one new module responsible for the cache. It will create a cache from the DB which will be used by the web. Also created a class to read from the cache.
Next thing would be to actually make parser run the cache creation after it will finish writing to the DB.
The other solution would be for the parser to actually create the cache while it is passing data to the DB - let us discuss it