Changes in the GUI
I have made various changes so that now we can switch between the views of Publio and Empik. For now the values are hardcoded, and if another library will be added I will need to make$
more generic based on the currently showed table.
This works now for presentation purposes and will be changed in final product.
This branch can be merged for now.